Refactoring which leads to adding /recherche/carte route
